Added hourly data
Items left to implement:
1) Digital Clock
2) RecycleView mechanics fix

+++ app/src/main/java/com/example/exam/data/retrofit/model/ForecastAPIResponse.kt
@@ -86,7 +86,7 @@ data class Current(
 	val visibility: Int,
 
 	@field:SerializedName("uvi")
-	val uvi: Int,
+	val uvi: Double,
 
 	@field:SerializedName("pressure")
 	val pressure: Int,
