+++ app/src/main/java/com/example/exam/model/InnerRecyclerWeatherItem.kt
@@ -1,3 +1,3 @@
 package com.example.exam.model
 
-data class InnerRecyclerWeatherItem(val hour: String, val tempInCelcius: Float)
+data class InnerRecyclerWeatherItem(val hour: String, val tempInCelcius: Double)
